Refactored code and moved out common components.
