+++ webapp/src/utils/validation/contentHelpers.js
@@ -1,0 +1,8 @@
+export const titleOptions = [
+  { value: "Mr", label: "Mr" },
+  { value: "Mrs", label: "Mrs" },
+  { value: "Ms", label: "Ms" },
+  { value: "Hon", label: "Hon" },
+  { value: "Prof", label: "Prof" },
+  { value: "Dr", label: "Dr" }
+];
